[Server] Fix some issues with K8s version
- Create specs.yaml file in init container if not exists
- Fix permissions of Python packages in auth image
- Mount application.yml into proxy container

+++ server/proxy/boot.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
-cp data/application.yml .
+[[ -e data/application.yml ]] && cp data/application.yml .
 [[ -e data/logo.png ]] && cp data/logo.png templates/2col/assets/img/logo.png
 exec java -jar shinyproxy.jar
